agents/alfred/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==15.0.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
docstring-to-markdown==0.17
docstring_parser==0.16
eval_type_backport==0.2.2
fastapi==0.115.14
flatbuffers==25.2.10
frozenlist==1.7.0
groq==0.29.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
janus==2.0.0
jedi==0.19.2
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
livekit==1.0.11
livekit-agents==1.1.5
livekit-api==1.0.3
livekit-plugins-cartesia==1.1.5
livekit-plugins-deepgram==1.1.5
livekit-plugins-groq==1.1.5
livekit-plugins-openai==1.1.5
livekit-plugins-silero==1.1.5
livekit-protocol==1.0.4
mcp==1.10.1
mpmath==1.3.0
multidict==6.6.3
mypy==1.16.1
mypy_extensions==1.1.0
nats-py==2.10.0
nest-asyncio==1.6.0
numpy==2.3.1
onnxruntime==1.22.0
openai==1.93.0
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-asyncio==1.0.0
python-dotenv==1.1.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.2
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
scipy==1.16.0
sniffio==1.3.1
sounddevice==0.5.2
sse-starlette==2.3.6
starlette==0.46.2
structlog==25.4.0
sympy==1.14.0
tqdm==4.67.1
types-protobuf==4.25.0.20240417
types-PyYAML==6.0.12.20250516
typing-inspection==0.4.1
typing_extensions==4.14.1
ujson==5.10.0
urllib3==2.5.0
uv==0.7.19
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
